category permission added

+++ app/Http/Controllers/CategoryController.php
@@ -7,6 +7,11 @@ use App\Models\Category;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,10 +19,14 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::all();
-        return view('category.index')->with([
-            'categories' => $categories
-        ]);
+        if(auth()->user()->can('index category')){
+            $categories = Category::all();
+            return view('category.index')->with([
+                'categories' => $categories
+            ]);
+        }else{
+            return redirect('home')->with('error','Unauthorized Access');
+        }
     }
 
     /**
@@ -27,7 +36,12 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('category.create');
+        if(auth()->user()->can('create category')){
+            return view('category.create');
+        }else{
+            return redirect('home')->with('error','Unauthorized Access');
+        }
+
     }
 
     /**
@@ -38,25 +52,28 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-           'title' => 'required|max:20',
-           'short_code' => 'required|max:5'
-        ]);
-        $category = new Category;
-        $category->title = $request->title;
-        $category->short_code = $request->short_code;
-        if($request->has('description')){
-            $category->description = $request->description;
-        }
-        $category->is_active = $request->has('is_active');
-        $category->is_featured = $request->has('is_featured');
+        if(auth()->user()->can('store category')){
+            $request->validate([
+                'title' => 'required|max:20',
+                'short_code' => 'required|max:5'
+            ]);
+            $category = new Category;
+            $category->title = $request->title;
+            $category->short_code = $request->short_code;
+            if($request->has('description')){
+                $category->description = $request->description;
+            }
+            $category->is_active = $request->has('is_active');
+            $category->is_featured = $request->has('is_featured');
 
-        try{
-            $category->save();
-            return redirect(route('category.index'))->with('success','successfully stored');
-        }catch (\Exception $e){
-            return redirect()->back()->withErrors($e);
+            try{
+                $category->save();
+                return redirect(route('category.index'))->with('success','successfully stored');
+            }catch (\Exception $e){
+                return redirect()->back()->withErrors($e);
+            }
         }
+
     }
 
     /**
@@ -67,17 +84,20 @@ class CategoryController extends Controller
      */
     public function show($id)
     {
-        if(is_numeric($id)){
-            $category = Category::find($id);
-            if($category == null){
-                return redirect()->back()->with('error','category not exists!');
+        if(auth()->user()->can('show category')){
+            if(is_numeric($id)){
+                $category = Category::find($id);
+                if($category == null){
+                    return redirect()->back()->with('error','category not exists!');
+                }
+                return view('category.show')->with([
+                    'category' => $category
+                ]);
+            }else{
+                return redirect()->back()->with('error','wrong url!');
             }
-            return view('category.show')->with([
-                'category' => $category
-            ]);
-        }else{
-            return redirect()->back()->with('error','wrong url!');
         }
+        return redirect('home')->with('error','Unauthorized Access!');
     }
 
     /**
@@ -88,18 +108,20 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
-        if(is_numeric($id)){
-            $category = Category::find($id);
-            if($category == null){
-                return redirect()->back()->with('error','category not exists!');
+        if(auth()->user()->can('edit category')){
+            if(is_numeric($id)){
+                $category = Category::find($id);
+                if($category == null){
+                    return redirect()->back()->with('error','category not exists!');
+                }
+                return view('category.edit')->with([
+                    'category' => $category
+                ]);
+            }else{
+                return redirect()->back()->with('error','wrong url!');
             }
-            return view('category.edit')->with([
-                'category' => $category
-            ]);
-        }else{
-            return redirect()->back()->with('error','wrong url!');
         }
-
+        return redirect('home')->with('error','Unauthorized Access!');
     }
 
     /**
@@ -111,32 +133,35 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if(is_numeric($id)){
-            $category = Category::find($id);
-            if($category == null){
-                return redirect()->back()->with('error','category not exists!');
-            }
-            $request->validate([
-                'title' => 'required|max:20',
-                'short_code' => 'required|max:5'
-            ]);
-            $category->title = $request->title;
-            $category->short_code = $request->short_code;
-            if($request->has('description')){
-                $category->description = $request->description;
-            }
-            $category->is_active = $request->has('is_active');
-            $category->is_featured = $request->has('is_featured');
+        if(auth()->user()->can('update category')){
+            if(is_numeric($id)){
+                $category = Category::find($id);
+                if($category == null){
+                    return redirect()->back()->with('error','category not exists!');
+                }
+                $request->validate([
+                    'title' => 'required|max:20',
+                    'short_code' => 'required|max:5'
+                ]);
+                $category->title = $request->title;
+                $category->short_code = $request->short_code;
+                if($request->has('description')){
+                    $category->description = $request->description;
+                }
+                $category->is_active = $request->has('is_active');
+                $category->is_featured = $request->has('is_featured');
 
-            try{
-                $category->save();
-                return redirect(route('category.index'))->with('success','successfully updated!');
-            }catch (\Exception $e){
-                return redirect()->back()->withErrors($e);
+                try{
+                    $category->save();
+                    return redirect(route('category.index'))->with('success','successfully updated!');
+                }catch (\Exception $e){
+                    return redirect()->back()->withErrors($e);
+                }
+            }else{
+                return redirect()->back()->with('error','wrong url!');
             }
-        }else{
-            return redirect()->back()->with('error','wrong url!');
         }
+        return redirect('home')->with('error','Unauthorized Access!');
     }
 
     /**
@@ -147,19 +172,22 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        if(is_numeric($id)){
-            $category = Category::find($id);
-            if($category == null){
-                return redirect()->back()->with('error','category not exists!');
+        if(auth()->user()->can('delete category')){
+            if(is_numeric($id)){
+                $category = Category::find($id);
+                if($category == null){
+                    return redirect()->back()->with('error','category not exists!');
+                }
+                try{
+                    $category->delete();
+                    return redirect(route('category.index'))->with('success','successfully deleted!');
+                }catch (\Exception $e){
+                    return redirect()->back()->withErrors($e);
+                }
+            }else{
+                return redirect()->back()->with('error','wrong url!');
             }
-            try{
-                $category->delete();
-                return redirect(route('category.index'))->with('success','successfully deleted!');
-            }catch (\Exception $e){
-                return redirect()->back()->withErrors($e);
-            }
-        }else{
-            return redirect()->back()->with('error','wrong url!');
         }
+        return redirect('home')->with('error','Unauthorized Access!');
     }
 }
